CA-227062: remove spurious backtrace
When calling `xe host-*` with the `host` param we were first checking if host represented a `uuid` using `host.get_by_uuid`, and only after a failure we were probing for the hostname (using the `name-label` field for the search). This was leaving a spurious backtrace in the logs when `host` was actually the hostname, and could potentially fail when in case the `name-label` was modified.

The fix involves replacing the host param with `name-label` and doing the search with them first, falling back to a search for `uuid` only in case of failure.

The potential issue with the hostname lookup is still open, but a
warning comment has been added to the code.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>