Initial commit of my performance fix to VBD unplug, to be tested
The existing code attempted to lock every VDI on the SR (on every unplug operation), and this led to operations getting stuck waiting on locks. The optimisation instead checks the datapaths of each VDI without locking, and only locks and performs operations on those VDIs that contain the datapath being removed.

This relies on the assumption that no VDIs with the datapath will become present between this check and the end of the destroy datapath call. I've added code to check for this race condition, to ensure that it doesn't occur in practice through the tests.

I've also simplified the code based on the further assumption (also tested by this code) that a datapath can only be attached to zero or one VDIs in an SR. Therefore we don't need to deal with a list of VDIs but simply an option of zero or one - and this follows through the destroy_sr call.

There's more code here than is needed for the fix because of these checks, so it can be simplified further once the assumptions are proven.